Add files via upload
Written by Brian Lesko
Multiple Linear Regression written in Python Jupyter Notebook for ECE 5307. In this ML Lab, the joint current is predicted using 7 features. The contents of this lab focus on the use of multiple linear regression to calibrate robot control. The lab uses data from the TU Dortmund's Multi-Elastic-Link Robot Identification Dataset, specifically, the current draw into one of the joints as a function of the robot motion. The lab starts by loading and visualizing the data, then extracting the time indices and current into the second joint into vectors. The lab then constructs the training data using all the samples from the experiment 1 dataset, and fits a linear model using the sklearn.linear_model module. The lab then trains the model on the training data, and calculates the R^2 coefficient-of-determination on the training data.